hooksにサインイン関係の処理を移行した
